requirements.txt
praw
python-dotenv
langdetect
google-api-python-client
python-dateutil
emoji
contractions